remove_git_lfs_and_work_with_new_structure
