--- .gitattributes
@@ -1,1 +1,0 @@
-*.pkl filter=lfs diff=lfs merge=lfs -text
