--- movie-recommender-system-tmdb-dataset-main/README.md
@@ -1,2 +1,0 @@
-# movie-recommender-system-tmdb-dataset
-A content based movie recommender system using cosine similarity
