--- movie-recommender-system-tmdb-dataset-main/movie_list.pkl
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e29eeb751a7a357a3231bed89cda41e20eaaf1510890b93843d290835323ce49
-size 2240181
