--- movie-recommender-system-tmdb-dataset-main/similarity.pkl
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5c919f931639c87771eaa3d2a0e1ed6c5492a646b1a7db40ff5903349d85cd07
-size 184781251
